Refactored some object and bucket code

+++ examples/s3.rs
@@ -21,7 +21,8 @@
 #![allow(unused_variables)]  // Mainly got tired of looking at warnings so added this :)
 
 extern crate aws_sdk_rust;
-#[macro_use] extern crate lsio;
+#[macro_use]
+extern crate lsio;
 extern crate url;
 //extern crate hyper;
 extern crate rustc_serialize;
@@ -87,13 +88,15 @@ fn main() {
 
     // For cli version see s3lsio cli
     let bucket_name: &str = "cm2test6";
+    let width: usize = 120;
 
     // NOTE: repeat_color and println_color are macros from the lsio library
 
-    repeat_color!(term::color::GREEN, "=", 80);
+    repeat_color!(term::color::GREEN, "=", "Start", width);
 
 
-    println!("---------------------------create_bucket----------------------------------------");
+    repeat_color_with_ends!(term::color::WHITE, "-", "create_bucket", "", "", width);
+
     let mut bucket = CreateBucketRequest::default();
     bucket.bucket = bucket_name.to_string();
 
@@ -103,7 +106,8 @@ fn main() {
     }
 
 
-    println!("---------------------------get_bucket_notification_configuration----------------");
+    repeat_color!(term::color::WHITE, "-", "get_bucket_notification_configuration", width);
+
     let mut bucket_notify = GetBucketNotificationConfigurationRequest::default();
     bucket_notify.bucket = bucket_name.to_string();
 
@@ -113,7 +117,8 @@ fn main() {
     }
 
 
-    println!("---------------------------get_bucket_logging----------------------------------");
+    repeat_color!(term::color::WHITE, "-", "get_bucket_logging", width);
+
     let mut bucket_logging = GetBucketLoggingRequest::default();
     bucket_logging.bucket = bucket_name.to_string();
 
@@ -123,7 +128,8 @@ fn main() {
     }
 
 
-    println!("---------------------------put_bucket_acl-------------------------------------");
+    repeat_color!(term::color::WHITE, "-", "put_bucket_acl", width);
+
     let mut put_bucket_acl = PutBucketAclRequest::default();
     put_bucket_acl.bucket = bucket_name.to_string();
     put_bucket_acl.acl = Some(CannedAcl::PublicRead);
@@ -134,7 +140,8 @@ fn main() {
     }
 
 
-    println!("---------------------------get_bucket_acl------------------------------------");
+    repeat_color!(term::color::WHITE, "-", "get_bucket_acl", width);
+
     let mut get_bucket_acl = GetBucketAclRequest::default();
     get_bucket_acl.bucket = bucket_name.to_string();
 
@@ -144,7 +151,8 @@ fn main() {
     }
 
 
-    println!("---------------------------put_object---------------------------------------");
+    repeat_color!(term::color::WHITE, "-", "put_object", width);
+
     let mut put_object = PutObjectRequest::default();
     put_object.bucket = bucket_name.to_string();
     put_object.key = "mytest.txt".to_string();
@@ -156,7 +164,8 @@ fn main() {
     }
 
 
-    println!("---------------------------get_object---------------------------------------");
+    repeat_color!(term::color::WHITE, "-", "get_object", width);
+
     let mut get_object = GetObjectRequest::default();
     get_object.bucket = bucket_name.to_string();
     get_object.key = "mytest.txt".to_string();
@@ -167,12 +176,17 @@ fn main() {
     }
 
 
-    println!("---------------------------create_multipart_upload------------------------------");
+    repeat_color!(term::color::WHITE, "-", "create_multipart_upload", width);
+
     // This is the first thing that needs to be done. Initiate the request and get the uploadid.
     // Generate a test file of 8MB in size...
     let test_abort: bool = false;
     let file_size: u16 = 8;
+    // NOTE: .gitignore has this file name as an ignore. If you change this file name then change .gitignore if you also want to issue a PR.
     let file_name: &str = "test.multipart.upload.file";
+    // NOTE: The temp file will be removed after the test. If you want to keep the file then set this to false.
+    let file_remove: bool = true;
+    let file_create: bool = true;
 
     let mut create_multipart_upload_output: Option<CreateMultipartUploadOutput> = None;
     let mut create_multipart_upload = CreateMultipartUploadRequest::default();
@@ -184,12 +198,15 @@ fn main() {
             println_color!(term::color::GREEN, "{:#?}", output);
             create_multipart_upload_output = Some(output);
             // Only for *nix based systems - the following command
-            let result = run_cli(format!("dd if=/dev/zero ibs={}m count=1 of={}", file_size, file_name.to_string()));
+            if file_create {
+                let result = run_cli(format!("dd if=/dev/zero ibs={}m count=1 of={}", file_size, file_name.to_string()));
+            }
         },
         Err(e) => println_color!(term::color::RED, "{:#?}", e),
     }
 
-    println!("---------------------------upload_part------------------------------------------");
+    repeat_color!(term::color::WHITE, "-", "upload_part", width);
+
     // You will need to decide on your upload part sizes. The minimum size is 5MB per chunk exept
     // the last one with a maximum size of 5GB per part. The total file size can't exceed 5TB.
     // Also, you can break the parts up into at most 10,000 parts. You want to split your file
@@ -210,7 +227,7 @@ fn main() {
         let upload_id: &str = &create_multipart_upload.upload_id;
         let mut parts_list: Vec<String> = Vec::new();
 
-        println!("---------------------------part-1-----------------------------------------------");
+        repeat_color!(term::color::WHITE, "-", "part-1", width);
 
         // Create read buffer for bytes and read in first part.
         //let path = Path::new(file_name);
@@ -221,7 +238,6 @@ fn main() {
         let metadata = f1.metadata().unwrap();
         let min_size: u64 = 5242880;
         let len: usize = (metadata.len() - min_size) as usize;
-        println!("{:?}", len);
         // NOTE: Don't do this in a dynamic envrionment since the metadata.len() is u64 and Vec can't handle that size.
         let mut part1_buffer: Vec<u8> = Vec::with_capacity(min_size as usize); // 5MB
         let mut part2_buffer: Vec<u8> = Vec::with_capacity(if len > min_size as usize {min_size as usize} else {len});
@@ -257,7 +273,8 @@ fn main() {
         // can break it up into 5MB and 3MB for two parts. Could put in a loop and make it more
         // useful.
 
-        println!("---------------------------list_multipart_uploads-------------------------------");
+        repeat_color!(term::color::WHITE, "-", "list_multipart_uploads", width);
+
         let mut list_multipart_uploads = ListMultipartUploadsRequest::default();
         list_multipart_uploads.bucket = bucket_name.to_string();
 
@@ -266,7 +283,8 @@ fn main() {
             Err(e) => println_color!(term::color::RED, "{:#?}", e),
         }
 
-        println!("---------------------------list_parts-------------------------------------------");
+        repeat_color!(term::color::WHITE, "-", "list_parts (#1)", width);
+
         let mut list_parts = ListPartsRequest::default();
         list_parts.bucket = bucket_name.to_string();
         list_parts.upload_id = upload_id.to_string();
@@ -279,7 +297,7 @@ fn main() {
 
         // If test_abort is true then don't upload the last part so that we can test abort.
         if !test_abort {
-            println!("------------------------part-2-----------------------------------------------");
+            repeat_color!(term::color::WHITE, "-", "part-2", width);
 
             let seek_result = f2.seek(SeekFrom::Start(min_size + 1)).unwrap();
 
@@ -304,7 +322,8 @@ fn main() {
             }
 
             // Just to show both parts now.
-            println!("------------------------list_parts-------------------------------------------");
+            repeat_color!(term::color::WHITE, "-", "list_parts (#2)", width);
+
             let mut list_parts = ListPartsRequest::default();
             list_parts.bucket = bucket_name.to_string();
             list_parts.upload_id = upload_id.to_string();
@@ -318,7 +337,8 @@ fn main() {
 
         // If the test_abort is true then abort the process.
         if test_abort {
-            println!("------------------------abort_upload-----------------------------------------");
+            repeat_color!(term::color::WHITE, "-", "abort_upload", width);
+
             let mut abort_multipart_upload = AbortMultipartUploadRequest::default();
             abort_multipart_upload.bucket = bucket_name.to_string();
             abort_multipart_upload.upload_id = upload_id.to_string();
@@ -330,7 +350,8 @@ fn main() {
             }
         } else {
             // Test complete if not testing abort
-            println!("------------------------complete_multipart_upload----------------------------");
+            repeat_color!(term::color::WHITE, "-", "complete_multipart_upload", width);
+
             let item_list : Vec<u8>;
 
             let mut complete_multipart_upload = CompleteMultipartUploadRequest::default();
@@ -354,10 +375,16 @@ fn main() {
                 Err(e) => println_color!(term::color::RED, "{:#?}", e),
             }
         }
+
+        // Remove temp file - ignore result
+        if file_remove {
+            let result = run_cli(format!("rm -f {}", file_name.to_string()));
+        }
     }
 
 
-    println!("---------------------------put_object_acl-----------------------------------");
+    repeat_color!(term::color::WHITE, "-", "put_object_acl", width);
+
     let mut put_object_acl = PutObjectAclRequest::default();
     put_object_acl.bucket = bucket_name.to_string();
     put_object_acl.key = "mytest.txt".to_string();
@@ -369,7 +396,8 @@ fn main() {
     }
 
 
-    println!("--------------------------get_object_acl-----------------------------------");
+    repeat_color!(term::color::WHITE, "-", "get_object_acl", width);
+
     let mut get_object_acl = GetObjectAclRequest::default();
     get_object_acl.bucket = bucket_name.to_string();
     get_object_acl.key = "mytest.txt".to_string();
@@ -384,7 +412,8 @@ fn main() {
     }
 
 
-    println!("--------------------------head_bucket--------------------------------------");
+    repeat_color!(term::color::WHITE, "-", "head_bucket", width);
+
     let bucket_head = HeadBucketRequest { bucket: bucket_name.to_string() };
 
     match client.head_bucket(&bucket_head) {
@@ -393,7 +422,8 @@ fn main() {
     }
 
 
-    println!("--------------------------put_bucket_versioning----------------------------");
+    repeat_color!(term::color::WHITE, "-", "put_bucket_versioning", width);
+
     let bucket_versioning = PutBucketVersioningRequest {
         bucket: bucket_name.to_string(),
         versioning_configuration: VersioningConfiguration {
@@ -410,7 +440,8 @@ fn main() {
     }
 
 
-    println!("--------------------------get_bucket_versioning--------------------------");
+    repeat_color!(term::color::WHITE, "-", "get_bucket_versioning", width);
+
     let bucket_versioning = GetBucketVersioningRequest { bucket: bucket_name.to_string() };
 
     match client.get_bucket_versioning(&bucket_versioning) {
@@ -419,7 +450,8 @@ fn main() {
     }
 
 
-    println!("--------------------------delete_object---------------------------------");
+    repeat_color!(term::color::WHITE, "-", "delete_object", width);
+
     let mut del_object = DeleteObjectRequest::default();
     del_object.bucket = bucket_name.to_string();
     del_object.key = "mytest.txt".to_string();
@@ -430,7 +462,8 @@ fn main() {
     }
 
 
-    println!("---------------------------list_object_versions-------------------------");
+    repeat_color!(term::color::WHITE, "-", "list_object_versions", width);
+
     let mut bucket_versioning = ListObjectVersionsRequest::default();
     bucket_versioning.bucket = bucket_name.to_string();
 
@@ -440,10 +473,12 @@ fn main() {
     }
 
 
-    println!("---------------------------list_objects--------------------------------");
+    repeat_color!(term::color::WHITE, "-", "list_objects (again)", width);
+
     let mut list_objects = ListObjectsRequest::default();
     list_objects.bucket = bucket_name.to_string();
-    // list_objects.version = Some(2);
+    // NOTE: The default is version 1 listing. You must set it to version 2.
+    list_objects.version = Some(2);
 
     match client.list_objects(&list_objects) {
         Ok(objects) => {
@@ -458,15 +493,29 @@ fn main() {
     }
 
 
-    println!("---------------------------delete_bucket-------------------------------");
+    repeat_color!(term::color::WHITE, "-", "delete_bucket", width);
+
     let bucket = DeleteBucketRequest { bucket: bucket_name.to_string() };
 
     match client.delete_bucket(&bucket) {
         Ok(bucket) => println_color!(term::color::GREEN, "{:#?}", bucket),
-        Err(e) => println_color!(term::color::RED, "{:#?}", e),
+        Err(e) => {
+            println_color!(term::color::RED, "{:#?}", e);
+
+            repeat_color!(term::color::WHITE, "-", "list_object_versions", width);
+
+            let mut bucket_versioning = ListObjectVersionsRequest::default();
+            bucket_versioning.bucket = bucket_name.to_string();
+
+            match client.list_object_versions(&bucket_versioning) {
+                Ok(version) => println_color!(term::color::GREEN, "{:#?}", version),
+                Err(e) => println_color!(term::color::RED, "{:#?}", e),
+            }
+        },
     }
 
-    println!("---------------------------list_buckets-------------------------------");
+    repeat_color!(term::color::WHITE, "-", "list_buckets", width);
+
     match client.list_buckets() {
         Ok(output) => {
             println_color!(term::color::GREEN, "{:#?}", output);
@@ -476,5 +525,5 @@ fn main() {
         },
     }
 
-    repeat_color!(term::color::GREEN, "=", 80);
+    repeat_color!(term::color::GREEN, "=", "Finished", width);
 }
